Added layout component.
