+++ client/src/components/codeEditor/CodeEditor.js
@@ -3,7 +3,15 @@ import { useHistory } from "react-router-dom";
 import { withRouter } from 'react-router'
 import { Controlled as CodeMirror } from 'react-codemirror2';
 import styles from './CodeEditor.css';
+import Grid from '@material-ui/core/Grid';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import Button from '@material-ui/core/Button';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
 
+// codemirror stuff that may or may not work properly
 import 'codemirror/lib/codemirror.css';
 import 'codemirror/theme/material.css';
 
@@ -30,6 +38,7 @@ class CodeEditor extends Component {
     }
 
     componentDidMount(){
+      document.title = this.state.title;
         document.addEventListener('keydown', this.handleHotKeys);
         this.getDocument(this.props.match.params.documentId);
     }
@@ -119,11 +128,11 @@ class CodeEditor extends Component {
             scrollbarStyle: null,
             lineWrapping: true,
             autoCloseBrackets: true
-          };
+        };
 
         return(
-            <div>
-                <div className={styles.headerText}>
+          <div>
+          <div className={styles.headerText}>
                 <h3>Room Generator</h3>
                 <p>{this.state.title}</p>
                 <p>{this.state.room}</p>
@@ -141,7 +150,7 @@ class CodeEditor extends Component {
                     this.setState({ code });
                   }}
                 />
-            </div>
+        </div>
         )
     }
 }
